refactor: extract Settings and wizard workflow presenters under 500 lines
Move Method / Extract Class: Settings CLI and skills orchestration into presenters;
Setup Wizard CLI/skills workflow into controllers; Migration Wizard operations into
a workflow controller. Windows keep shell/DI/static facades. No intended behavior change.
