+++ Packages/src/Editor/Presentation/Setup/SetupWizardCliWorkflowController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7a13fc2d47454255863d9f9f35f90719
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
